finish all CRUD for all entities , but the postmapping didn't work

+++ src/main/java/com/company/AlaaAdinWebsite/dao/FactoryOwnerRepository.java
@@ -3,7 +3,7 @@ package com.company.AlaaAdinWebsite.dao;
 import com.company.AlaaAdinWebsite.entity.FactoryOwner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FactoryOwnerRepository extends JpaRepository<FactoryOwner,Long> {
+public interface FactoryOwnerRepository extends JpaRepository<FactoryOwner,Integer> {
 
     // that's it ... no need to write any code LOL!
 }
