+++ src/main/java/com/company/AlaaAdinWebsite/dao/SubCategoryRepository.java
@@ -3,6 +3,6 @@ package com.company.AlaaAdinWebsite.dao;
 import com.company.AlaaAdinWebsite.entity.SubCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubCategoryRepository extends JpaRepository<SubCategory,Long> {
+public interface SubCategoryRepository extends JpaRepository<SubCategory,Integer> {
     //No need to Write code;
 }
