+++ src/main/java/com/company/AlaaAdinWebsite/service/ClientService.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface ClientService {
     List<Client> findAll();
-    Optional<Client> findById(Long theId);
+    Optional<Client> findById(int theId);
     Client save(Client client);
-    void deleteById(Long theId);
+    void deleteById(int theId);
 }
